Fix conftest for external testing and fix style check
