+++ MANIFEST.in
@@ -1,7 +1,6 @@
 include README.rst
 include CHANGES.rst
 include pyproject.toml
-include conftest.py
 
 recursive-include docs *
 recursive-include licenses *
